requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
fonttools==4.25.0
future==0.18.3
kiwisolver @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_e26jwrjf6j/croot/kiwisolver_1672387151391/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1661440102531/work
munkres==1.1.4
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1649806371302/work
opencv-python==4.3.0.38
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
Pillow==9.4.0
pyparsing @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_3a17y2delq/croots/recipe/pyparsing_1661452538853/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
scipy==1.5.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
torch==1.6.0
torchvision==0.7.0
tornado @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1fimz6o0gc/croots/recipe/tornado_1662061695695/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
